extract layout picker component
