+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/meal/ui/cardsettings/MealCardSettings.kt
@@ -1,13 +1,8 @@
 package com.maksimowiczm.foodyou.feature.meal.ui.cardsettings
 
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListScope
@@ -17,16 +12,13 @@ import androidx.compose.material3.LargeTopAppBar
 import androidx.compose.material3.ListItem
 import androidx.compose.material3.ListItemDefaults
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.hapticfeedback.HapticFeedbackType
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.platform.LocalHapticFeedback
@@ -99,12 +91,6 @@ internal fun MealCardSettings(
     modifier: Modifier = Modifier
 ) {
     val hapticFeedback = LocalHapticFeedback.current
-    val onLayoutChange: (MealCardsLayout) -> Unit = {
-        if (layout != it) {
-            hapticFeedback.performHapticFeedback(HapticFeedbackType.ToggleOn)
-            onLayoutChange(it)
-        }
-    }
 
     val scrollBehavior = TopAppBarDefaults.exitUntilCollapsedScrollBehavior()
 
@@ -125,40 +111,16 @@ internal fun MealCardSettings(
             contentPadding = paddingValues
         ) {
             item {
-                Row(
-                    modifier = Modifier.fillMaxWidth().padding(vertical = 8.dp),
-                    horizontalArrangement = Arrangement.SpaceEvenly
-                ) {
-                    LayoutContainer(
-                        onLayoutChange = { onLayoutChange(MealCardsLayout.Horizontal) }
-                    ) {
-                        LayoutHorizontal()
-                        Spacer(Modifier.height(8.dp))
-                        Text(
-                            text = stringResource(Res.string.headline_horizontal),
-                            style = MaterialTheme.typography.labelLarge
-                        )
-                        RadioButton(
-                            selected = layout == MealCardsLayout.Horizontal,
-                            onClick = { onLayoutChange(MealCardsLayout.Horizontal) }
-                        )
-                    }
-
-                    LayoutContainer(
-                        onLayoutChange = { onLayoutChange(MealCardsLayout.Vertical) }
-                    ) {
-                        LayoutVertical()
-                        Spacer(Modifier.height(8.dp))
-                        Text(
-                            text = stringResource(Res.string.headline_vertical),
-                            style = MaterialTheme.typography.labelLarge
-                        )
-                        RadioButton(
-                            selected = layout == MealCardsLayout.Vertical,
-                            onClick = { onLayoutChange(MealCardsLayout.Vertical) }
-                        )
-                    }
-                }
+                LayoutPicker(
+                    layout = layout,
+                    onLayoutChange = {
+                        if (layout != it) {
+                            hapticFeedback.performHapticFeedback(HapticFeedbackType.ToggleOn)
+                            onLayoutChange(it)
+                        }
+                    },
+                    modifier = Modifier.fillMaxWidth().padding(vertical = 8.dp)
+                )
             }
 
             item {
@@ -191,23 +153,6 @@ internal fun MealCardSettings(
                 )
             }
         }
-    }
-}
-
-@Composable
-private fun LayoutContainer(
-    onLayoutChange: () -> Unit,
-    modifier: Modifier = Modifier,
-    content: @Composable () -> Unit
-) {
-    Column(
-        modifier = modifier
-            .clip(MaterialTheme.shapes.medium)
-            .clickable { onLayoutChange() }
-            .padding(8.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        content()
     }
 }
